backend/requirements.txt
Django==4.2.13
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==24.2
PyMySQL==1.1.1
Pillow>=10.3.0
python-decouple==3.8
cryptography
gunicorn==21.2.0
whitenoise==6.7.0